+++ src/lib.rs
@@ -31,6 +31,7 @@ pub mod paste_transform;
 pub mod profile;
 pub mod profile_drawer_ui;
 pub mod profile_modal_ui;
+pub mod progress_bar;
 pub mod quit_confirmation_ui;
 pub mod renderer;
 pub mod scroll_state;